Group preview: choose suitable hit. Don't highlight list of relations.
We now fetch 10 hits and choose the first suitable hit for the group by
preview. A hit is suitable if it has otherFields set (i.e. a cross-field
match was found); if no such hit was fetched, the first hit is used. This
fixes the preview not showing target fields if the first hit doesn't have
a target hit (parallel corpora).

When highlighting matchInfo in the preview, we now skip the list-of-relations
matchInfo type. This matchInfo is not suitable for grouping on, because it
is a variable length list of relations between two spans, not a specific,
tagged relation that was queried.