Improve error message for 404 response from backend.
A 404 response may mean BlackLab Server isn't running or couldn't be found,
but might also mean you're trying to open a corpus that doesn't exist.

We now check if the response seems to be from BlackLab; if so, we print it.
If not, we indicate that there's probably a configuration problem that prevents
us from contacting BlackLab.