Enable mixed text direction corpora.
To create a corpus with both right-to-left and left-to-right
documents: set a corpus-wide default textDirection in your .blf.yaml
file. You can override this default for a document by indexing a
metadata field named 'textDirection'. It should get the value 'rtl'
or 'ltr'. Each hits should automatically be rendered in the correct
text direction.

Note that there may be lingering issues with this. Please file an
issue or create a PR if you find any problems.