Re-add banner message.
BannerMessage is exposed in websiteconfig now for ease of use clientside, still reads from globalconfig.

Update localstore utility to allow expiring keys. The banner uses this, it stores the banner whenever the user hides it, with a timeout of one week. If the banner message changes in the meantime, the banner will unhide. This is identical to the old behavior, but without the cookies.