Fix expanded hits not clearing state when paginating
This was due to vue reusing component instances due to the :key prop not being properly set. The :key prop should indicate that the component hasn't changed, but in reality the entire hit had changed, so the :key should be different. Since we can't righly construct a proper key without stringifying the entire hit (since we show pretty much all of it), simply remove the key altogether and make sure we don't hold over any data when the hit passed to a HitRow[Details] instance changes.