Use summary.pattern.otherFields to determine parallel fields to show.
Before we used targetFields from the query store, but that is only populated
if the different fields correspond to different UI elements. If the query cannot
be placed in separate UI elements, the entire BCQL query is placed in a single
Expert view box, and we need BLS' summary.pattern.otherFields to know the actual
parallel target fields.