Allow customizing sort/group options. Rename i18n/custom names.
Sort/group options can be customized per option group. The customize()
function will be called for each option group. It can inspect the option
group and customize based on the contents.
It might be useful to give each group a recognizable id so we know what
sort of options this group contains, i.e. metadata, span filters, etc.
Right now we can look at the label but this is an i18n'ed string.

Renaming: "within" was renamed to "spans", "withinAttributes" to
"spanAttributes", etc. because it is used for more than just the within
widget now.