Fix page state not restoring properly when using browser back button
This error was pretty rare, as we cache the state to avoid having to decode the url every time the user uses the back button, but when that cache is not available for any reason we fall back to decoding the url, which would fail.

The reason for that was that the url-state-parser has been changed to be async. Typescript didn't catch this, and were trying to paste a promise into our vuex store, which obviously didn't work.