One last fix to the debug checkbox & wide view
Wide view seemed to work fine in development mode, but in production, it ran before the components were fully mounted. Must be something to do with module evaluation order in debug vs prod, best not to rely on that behavior.
Instead, manually toggle the container/container-fluid class in the main search page component.

Also, debug checkbox should always be visible in debug mode, and in the current setup, setting debugInfo=false in config would hide the check even in dev mode, which isn't what I intended.