Attempt to fix the github runner crashing
There is an issue where the github CI used 'npm ci' to install locked versions of deps (good!).
However the package-lock.json could very well be generated on a windows machine, which would cause npm to pull in a windows-only binary for compiled dependencies.
Which would cause CI to fail...