Fix wrong relative group size when grouping docs by metadata
What's displayed in the table is complex:
There are 4 levels of data, from largest to smallest:
1. the total corpus, i.e. all docs
2. the total search space, i.e. all docs that match the Filters, e.g. "all docs from France"
3. the total group space, i.e. all docs that match the Filters + the current group's Metadata, e.g. "all docs from France [the filter], with year 2015 [grouped on year])
4. the actual group, i.e. all docs that match the Filters + the group's Metadata + that have at least one Hit that falls within this group. This is only available when searching for Hits and also grouping by at least one Hit property (e.g. Lemma)

The situation that was broken:
- searching without Pattern
- grouping on document metadata

We were comparing the side of 4. to the size of 3. Note that because there is no Pattern, thus no Hits, these two levels are equal in size, so all groups would always show as "100%" for the relative group size.
Edit this to compare 3. to 2. , which is actually what we wanted to show, e.g. "the size of all docs from 2015 (the group) from France (the Filter), compared to the size of all docs from France (the Filter)".