Fix incorrect handling of browser locale in i18n
The user agent sometimes returns a full language code e.g. 'nl-nl', but other times only the primary portion e.g. 'nl'. We did not account for this, so the short form would result in the locale not being found and a bunch of warnings being logged before defaulting to the english locale.
We now check if we have a partial match instead of an exact match, so detecting the user's preferred locale should be more robust.