Add dockerhub build hook to automatically determine correct blacklab base image
This hook dynamically sets the BLACKLAB_IMAGE_VERSION arg in the dockerfile used to build our image.
It does this by trying to determine the major version from the current branchname or git tag being built, with special behavior for 'dev' and 'main'/'latest'. If it fails to match, a warning is printed and 'dev' is used as BlackLab base image.